Double size of EntryClass
For the foot profile
[assert][140167020267392] /mnt/persist/api-osrm-routed-no-stxxl/2017-07-10-140458/osrm/src/util/guidance/entry_class.cpp:14
in: void osrm::util::guidance::EntryClass::activate(uint32_t): index < 8 * sizeof(FlagBaseType)

+++ include/util/guidance/entry_class.hpp
@@ -36,7 +36,7 @@ namespace guidance
 
 class EntryClass
 {
-    using FlagBaseType = std::uint32_t;
+    using FlagBaseType = std::uint64_t;
 
   public:
     constexpr EntryClass() : enabled_entries_flags(0) {}
